Remove browser-based metadata download (#3970)
* Remove browser-based metadata download

* Fix Javadoc, add TODO

* Further cleanup

* Remove leftover reference to DownloadService from UI

* License header and note of origin for DailyCheck

* Add Javadoc with note about the update frequency
