+++ core/src/main/java/hudson/model/DownloadService.java
@@ -33,7 +33,6 @@ import hudson.init.InitMilestone;
 import hudson.init.Initializer;
 import hudson.util.FormValidation;
 import hudson.util.FormValidation.Kind;
-import hudson.util.QuotedStringTokenizer;
 import hudson.util.TextFile;
 import static java.util.concurrent.TimeUnit.DAYS;
 import java.io.File;
@@ -46,20 +45,14 @@ import java.net.URLConnection;
 import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import jenkins.model.DownloadSettings;
 import jenkins.model.Jenkins;
 import net.sf.json.JSONException;
 import net.sf.json.JSONObject;
 import org.apache.commons.io.IOUtils;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
-import org.kohsuke.stapler.Stapler;
-import org.kohsuke.stapler.StaplerRequest;
-import org.kohsuke.stapler.StaplerResponse;
-import org.kohsuke.stapler.interceptor.RequirePOST;
 
 /**
  * Service for plugins to periodically retrieve update data files
@@ -72,7 +65,7 @@ import org.kohsuke.stapler.interceptor.RequirePOST;
  * @author Kohsuke Kawaguchi
  */
 @Extension
-public class DownloadService extends PageDecorator {
+public class DownloadService {
 
     /**
      * the prefix for the signature validator name
@@ -80,59 +73,12 @@ public class DownloadService extends PageDecorator {
     private static final String signatureValidatorPrefix = "downloadable";
     /**
      * Builds up an HTML fragment that starts all the download jobs.
+     *
+     * @deprecated browser-based download has been disabled
      */
+    @Deprecated
     public String generateFragment() {
-        if (!DownloadSettings.usePostBack()) {
-            return "";
-        }
-    	if (neverUpdate) return "";
-        if (doesNotSupportPostMessage())  return "";
-
-        StringBuilder buf = new StringBuilder();
-        if(Jenkins.get().hasPermission(Jenkins.READ)) {
-            long now = System.currentTimeMillis();
-            for (Downloadable d : Downloadable.all()) {
-                if(d.getDue()<now && d.lastAttempt+TimeUnit.SECONDS.toMillis(10)<now) {
-                    buf.append("<script>")
-                       .append("Behaviour.addLoadEvent(function() {")
-                       .append("  downloadService.download(")
-                       .append(QuotedStringTokenizer.quote(d.getId()))
-                       .append(',')
-                       .append(QuotedStringTokenizer.quote(mapHttps(d.getUrl())))
-                       .append(',')
-                       .append("{version:").append(QuotedStringTokenizer.quote(Jenkins.VERSION)).append('}')
-                       .append(',')
-                       .append(QuotedStringTokenizer.quote(Stapler.getCurrentRequest().getContextPath()+'/'+getUrl()+"/byId/"+d.getId()+"/postBack"))
-                       .append(',')
-                       .append("null);")
-                       .append("});")
-                       .append("</script>");
-                    d.lastAttempt = now;
-                }
-            }
-        }
-        return buf.toString();
-    }
-
-    private boolean doesNotSupportPostMessage() {
-        StaplerRequest req = Stapler.getCurrentRequest();
-        if (req==null)      return false;
-
-        String ua = req.getHeader("User-Agent");
-        if (ua==null)       return false;
-
-        // according to http://caniuse.com/#feat=x-doc-messaging, IE <=7 doesn't support pstMessage
-        // see http://www.useragentstring.com/pages/Internet%20Explorer/ for user agents
-
-        // we want to err on the cautious side here.
-        // Because of JENKINS-15105, we can't serve signed metadata from JSON, which means we need to be
-        // using a modern browser as a vehicle to request these data. This check is here to prevent Jenkins
-        // from using older browsers that are known not to support postMessage as the vehicle.
-        return ua.contains("Windows") && (ua.contains(" MSIE 5.") || ua.contains(" MSIE 6.") || ua.contains(" MSIE 7."));
-    }
-
-    private String mapHttps(String url) {
-        return url;
+        return "";
     }
 
     /**
@@ -148,7 +94,7 @@ public class DownloadService extends PageDecorator {
 
     /**
      * Loads JSON from a JSONP URL.
-     * Metadata for downloadables and update centers is offered in two formats, both designed for download from the browser (predating {@link DownloadSettings}):
+     * Metadata for downloadables and update centers is offered in two formats, both designed for download from the browser (a feature since removed):
      * HTML using {@code postMessage} for newer browsers, and JSONP as a fallback.
      * Confusingly, the JSONP files are given the {@code *.json} file extension, when they are really JavaScript and should be {@code *.js}.
      * This method extracts the JSON from a JSONP URL, since that is what we actually want when we download from the server.
@@ -368,24 +314,6 @@ public class DownloadService extends PageDecorator {
             return null;
         }
 
-        /**
-         * This is where the browser sends us the data. 
-         */
-        @RequirePOST
-        public void doPostBack(StaplerRequest req, StaplerResponse rsp) throws IOException {
-            DownloadSettings.checkPostBackAccess();
-            long dataTimestamp = System.currentTimeMillis();
-            due = dataTimestamp+getInterval();  // success or fail, don't try too often
-
-            String json = IOUtils.toString(req.getInputStream(),"UTF-8");
-            FormValidation e = load(json, dataTimestamp);
-            if (e.kind != Kind.OK) {
-                LOGGER.severe(e.renderHtml());
-                throw e;
-            }
-            rsp.setContentType("text/plain");  // So browser won't try to parse response
-        }
-
         private FormValidation load(String json, long dataTimestamp) throws IOException {
             TextFile df = getDataFile();
             df.write(json);
@@ -497,13 +425,12 @@ public class DownloadService extends PageDecorator {
                 SystemProperties.getLong(Downloadable.class.getName()+".defaultInterval", DAYS.toMillis(1));
     }
 
+    // TODO this was previously referenced in the browser-based download, but should probably be checked for the server-based download
     public static boolean neverUpdate = SystemProperties.getBoolean(DownloadService.class.getName()+".never");
 
     /**
      * May be used to temporarily disable signature checking on {@link DownloadService} and {@link UpdateCenter}.
      * Useful when upstream signatures are broken, such as due to expired certificates.
-     * Should only be used when {@link DownloadSettings#isUseBrowser};
-     * disabling signature checks for in-browser downloads is <em>very dangerous</em> as unprivileged users could submit spoofed metadata!
      */
     public static boolean signatureCheck = !SystemProperties.getBoolean(DownloadService.class.getName()+".noSignatureCheck");
 }
