+++ core/src/main/java/jenkins/telemetry/impl/SecuritySystemProperties.java
@@ -24,7 +24,6 @@
 package jenkins.telemetry.impl;
 
 import hudson.Extension;
-import jenkins.model.DownloadSettings;
 import jenkins.model.Jenkins;
 import jenkins.telemetry.Telemetry;
 import jenkins.util.SystemProperties;
@@ -92,9 +91,6 @@ public class SecuritySystemProperties extends Telemetry {
         putBoolean(security, "jenkins.security.ClassFilterImpl.SUPPRESS_ALL", false);
         putBoolean(security, "org.kohsuke.stapler.Facet.allowViewNamePathTraversal", false);
         putBoolean(security, "org.kohsuke.stapler.jelly.CustomJellyContext.escapeByDefault", true);
-
-        // not controlled by a system property for historical reasons only
-        security.put("jenkins.model.DownloadSettings.useBrowser", Boolean.toString(DownloadSettings.get().isUseBrowser()));
 
         putStringInfo(security, "hudson.model.ParametersAction.safeParameters");
         putStringInfo(security, "hudson.model.DirectoryBrowserSupport.CSP");
