+++ core/src/main/resources/hudson/PluginManager/check.jelly
@@ -24,9 +24,8 @@ THE SOFTWARE.
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
 
-  <j:invokeStatic var="ds" className="jenkins.model.DownloadSettings" method="get"/>
   ${%lastUpdated(app.updateCenter.lastUpdatedString)}
-  <f:link href="${ds.useBrowser ? 'checkUpdates' : 'checkUpdatesServer'}" post="true" clazz="yui-button yui-submit-button submit-button primary">
+  <f:link href="checkUpdatesServer" post="true" clazz="yui-button yui-submit-button submit-button primary">
     <button>${%Check now}</button>
   </f:link>
   <j:if test="${app.pluginManager.lastErrorCheckUpdateCenters != null}">
