--- core/src/main/resources/hudson/model/UpdateCenter/PageDecoratorImpl/footer.jelly
@@ -1,51 +1,0 @@
-<!--
-The MIT License
-
-Copyright (c) 2004-2009, Sun Microsystems, Inc., Kohsuke Kawaguchi, Yahoo! Inc. 
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
--->
-
-<!--
-  If necessary, ask the client to fetch update-center data file for us.
-  This allows us to avoid proxy related configuration on the server,
-  and instead piggy-back on user connection.
-
-  This file is pulled into the layout.jelly
--->
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-  <j:invokeStatic var="enabled" className="jenkins.model.DownloadSettings" method="usePostBack"/>
-  <j:if test="${enabled}">
-  <j:forEach var="site" items="${app.updateCenter.sites}">
-    <j:if test="${site.due or forcedUpdateCheck}">
-      <script>
-        Behaviour.addLoadEvent(function() {
-            downloadService.download(
-                "${site.id}",
-                "${site.downloadUrl}",
-                {version:"${h.version}"},
-                "${rootURL}/updateCenter/byId/${site.id}/postBack",
-                null);
-        });
-      </script>
-    </j:if>
-  </j:forEach>
-  </j:if>
-</j:jelly>
