--- core/src/main/resources/jenkins/model/DownloadSettings/config.groovy
@@ -1,9 +1,0 @@
-package jenkins.security.DownloadSettings
-
-def f = namespace(lib.FormTagLib)
-
-f.section(title:_("Plugin Manager")) {
-    f.entry() {
-        f.checkbox(field: "useBrowser", title: _("Use browser for metadata download"))
-    }
-}
