--- core/src/main/resources/jenkins/model/DownloadSettings/help-useBrowser_bg.html
@@ -1,12 +1,0 @@
-<p>
-    Настройване браузърът, а не Jenkins, да изтегля метаданните (списъци с наличните
-    приставки, инструменти и т.н.).  Самото изтегляне на приставките и инструментите
-    ще се извърши от Jenkins, но така ще получите достъп поне да <em>разгледате</em>
-    обновените метаданни, когато Jenkins няма достъп до Internet, но вашия браузър
-    има (примерно защото ползва определен сървър-посредник, който е недостъпен за
-    Jenkins).
-</p>
-<p>
-    Използването на този режим не се препоръчва.  Той не е надежден и е ограничен
-    единствено до администраторите, защото само те могат да се възползват от него.
-</p>
