--- core/src/main/resources/jenkins/model/DownloadSettings/help-useBrowser_it.html
@@ -1,16 +1,0 @@
-<p>
-    Selezionare quest'opzione per forzare il browser dell'utente a scaricare i
-    metadati (elenchi dei plugin/strumenti disponibili, ecc.) anziché lasciar
-    eseguire tale operazione a Jenkins stesso.
-    I download dei file veri e propri (plugin, strumenti) continueranno a
-    essere operati da Jenkins stesso, ma quest'opzione può essere utilizzata
-    almeno per <em>visualizzare</em> nuovi metadati quando Jenkins non può
-    accedere a Internet (ma il proprio browser è in grado di farlo, forse
-    utilizzando qualche proxy speciale che Jenkins non è configurato per
-    usare).
-</p>
-<p>
-    L'utilizzo della modalità browser non è raccomandato.
-    Ha problemi di robustezza ed è stato limitato agli utenti con il privilegio
-    Globale/Amministra.
-</p>
