--- core/src/main/resources/jenkins/model/DownloadSettings/help-useBrowser_ja.html
@@ -1,7 +1,0 @@
-<div>
-    チェックすると、Jenkinsではなく使用しているブラウザで、使用可能なプラグインやツールなどのメタ情報をダウンロードします。
-    プラグインやツールのファイルはjenkins自身がダウンロードしますが、
-    ブラウザを使用することで、Jenkinsがインターネットにアクセスできないときでも、少なくとも新規のメタ情報を参照することができます
-    (おそらく、ある特別なプロキシーを使用していて、Jenkinsがそれを使用するように設定できなくても、
-    ブラウザは設定できるでしょう)。
-</div>
