+++ core/src/main/resources/jenkins/model/Messages_bg.properties
@@ -111,9 +111,6 @@ CLI.disable-job.shortDescription=\
  \u0418\u0437\u043a\u043b\u044e\u0447\u0432\u0430\u043d\u0435 \u043d\u0430 \u0437\u0430\u0434\u0430\u0447\u0430.
 CLI.enable-job.shortDescription=\
  \u0412\u043a\u043b\u044e\u0447\u0432\u0430\u043d\u0435 \u043d\u0430 \u0437\u0430\u0434\u0430\u0447\u0430.
-# Browser-based metadata download
-DownloadSettings.Warning.DisplayName=\
- \u0418\u0437\u0442\u0435\u0433\u043b\u044f\u043d\u0430 \u043d\u0430 \u043c\u0435\u0442\u0430\u0434\u0430\u043d\u043d\u0438\u0442\u0435 \u0447\u0440\u0435\u0437 \u0431\u0440\u0430\u0443\u0437\u044a\u0440
 # Enforce JNLP Slave Agent Port
 EnforceSlaveAgentPortAdministrativeMonitor.displayName=\
  \u041a\u043e\u043d\u043a\u0440\u0435\u0442\u0435\u043d \u043f\u043e\u0440\u0442 \u0437\u0430 JNLP \u043d\u0430 \u043f\u043e\u0434\u0447\u0438\u043d\u0435\u043d\u0438\u044f \u043a\u043e\u043c\u043f\u044e\u0442\u044a\u0440
