+++ test/src/test/java/hudson/model/DownloadServiceTest.java
@@ -1,7 +1,6 @@
 package hudson.model;
 
 import hudson.model.DownloadService.Downloadable;
-import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,54 +10,14 @@ import java.util.TreeSet;
 import hudson.tasks.Maven;
 import hudson.tools.DownloadFromUrlInstaller;
 import hudson.tools.ToolInstallation;
-import jenkins.model.DownloadSettings;
 import net.sf.json.JSONObject;
-import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.HudsonTestCase;
 import org.jvnet.hudson.test.WithoutJenkins;
-import org.kohsuke.stapler.StaplerResponse;
 
 /**
  * @author Kohsuke Kawaguchi
  */
 public class DownloadServiceTest extends HudsonTestCase {
-    private Downloadable job;
-
-    /**
-     * Makes sure that JavaScript on the client side for handling submission works.
-     */
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-        if (jenkins == null) {
-            return;
-        }
-        // this object receives the submission.
-        // to bypass the URL restriction, we'll trigger downloadService.download ourselves
-        job = new Downloadable("test", "UNUSED");
-        Downloadable.all().add(job);
-        DownloadSettings.get().setUseBrowser(true);
-    }
-
-    @Issue("JENKINS-5536")
-    public void testPost() throws Exception {
-        // we do not save with signature for toolInstallers,
-        //neither we check it in the getData method.
-
-        createWebClient().goTo("/self/testPost");
-        JSONObject d = job.getData();
-        assertEquals(hashCode(),d.getInt("hello"));
-
-        // TODO: test with a signature
-    }
-
-    /**
-     * This is where the browser should hit to retrieve data.
-     */
-    public void doData(StaplerResponse rsp) throws IOException {
-        rsp.setContentType("application/javascript");
-        rsp.getWriter().println("downloadService.post('test',{'hello':"+hashCode()+"})");
-    }
 
     @WithoutJenkins // could have been in core/src/test/ but update-center.json was already in test/src/test/ (used by UpdateSiteTest)
     public void testLoadJSON() throws Exception {
